+++ fastly/block_fastly_service_dynamicsnippet.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-	gofastly "github.com/fastly/go-fastly/v16/fastly"
+	gofastly "github.com/fastly/go-fastly/v17/fastly"
 )
 
 // DynamicSnippetServiceAttributeHandler provides a base implementation for ServiceAttributeDefinition.