+++ fastly/block_fastly_service_logging_bigquery.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
-	gofastly "github.com/fastly/go-fastly/v16/fastly"
+	gofastly "github.com/fastly/go-fastly/v17/fastly"
 )
 
 // BigQueryLoggingServiceAttributeHandler provides a base implementation for ServiceAttributeDefinition.
@@ -225,7 +225,7 @@ func (h *BigQueryLoggingServiceAttributeHandler) Update(ctx context.Context, d *
 		opts.ResponseCondition = gofastly.ToPointer(v.(string))
 	}
 	if v, ok := modified["placement"]; ok {
-		opts.Placement = gofastly.ToPointer(v.(string))
+		opts.Placement = gofastly.NewNullable(v.(string))
 	}
 	if v, ok := modified["account_name"]; ok {
 		opts.AccountName = gofastly.ToPointer(v.(string))