+++ fastly/block_fastly_service_logging_cloudfiles.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
-	gofastly "github.com/fastly/go-fastly/v16/fastly"
+	gofastly "github.com/fastly/go-fastly/v17/fastly"
 )
 
 // CloudfilesServiceAttributeHandler provides a base implementation for ServiceAttributeDefinition.
@@ -217,7 +217,7 @@ func (h *CloudfilesServiceAttributeHandler) Update(ctx context.Context, d *schem
 		opts.Region = gofastly.ToPointer(v.(string))
 	}
 	if v, ok := modified["placement"]; ok {
-		opts.Placement = gofastly.ToPointer(v.(string))
+		opts.Placement = gofastly.NewNullable(v.(string))
 	}
 	if v, ok := modified["period"]; ok {
 		opts.Period = gofastly.ToPointer(v.(int))