+++ fastly/block_fastly_service_logging_logshuttle.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
-	gofastly "github.com/fastly/go-fastly/v16/fastly"
+	gofastly "github.com/fastly/go-fastly/v17/fastly"
 )
 
 // LogshuttleServiceAttributeHandler provides a base implementation for ServiceAttributeDefinition.
@@ -152,7 +152,7 @@ func (h *LogshuttleServiceAttributeHandler) Update(ctx context.Context, d *schem
 		opts.FormatVersion = gofastly.ToPointer(v.(int))
 	}
 	if v, ok := modified["placement"]; ok {
-		opts.Placement = gofastly.ToPointer(v.(string))
+		opts.Placement = gofastly.NewNullable(v.(string))
 	}
 	if v, ok := modified["response_condition"]; ok {
 		opts.ResponseCondition = gofastly.ToPointer(v.(string))