+++ fastly/block_fastly_service_logging_splunk.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
-	gofastly "github.com/fastly/go-fastly/v16/fastly"
+	gofastly "github.com/fastly/go-fastly/v17/fastly"
 )
 
 // SplunkServiceAttributeHandler provides a base implementation for ServiceAttributeDefinition.
@@ -221,7 +221,7 @@ func (h *SplunkServiceAttributeHandler) Update(ctx context.Context, d *schema.Re
 		opts.ResponseCondition = gofastly.ToPointer(v.(string))
 	}
 	if v, ok := modified["placement"]; ok {
-		opts.Placement = gofastly.ToPointer(v.(string))
+		opts.Placement = gofastly.NewNullable(v.(string))
 	}
 	if v, ok := modified["token"]; ok {
 		opts.Token = gofastly.ToPointer(v.(string))