+++ fastly/data_source_fastly_ngwaf_account_rules.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/fastly/terraform-provider-fastly/fastly/hashcode"
 
-	"github.com/fastly/go-fastly/v16/fastly/ngwaf/v1/rules"
-	"github.com/fastly/go-fastly/v16/fastly/ngwaf/v1/scope"
+	"github.com/fastly/go-fastly/v17/fastly/ngwaf/v1/rules"
+	"github.com/fastly/go-fastly/v17/fastly/ngwaf/v1/scope"
 )
 
 func dataSourceFastlyNGWAFAccountRules() *schema.Resource {